requirements.txt
wfdb==4.1.2
numpy>=1.24,<2.0
pandas>=2.0,<2.3
scipy>=1.10,<1.13
scikit-learn>=1.3,<1.6
xgboost>=2.0,<2.1
matplotlib>=3.7,<3.10
tqdm>=4.66
pyyaml>=6.0
joblib>=1.3
requests>=2.31
